+++ rdsa_utils/io/input.py
@@ -1,5 +1,6 @@
 """Module containing generic input functionality code."""
 
+import hashlib
 import json
 import logging
 from pathlib import Path
@@ -127,6 +128,27 @@ def file_size(
     """
     if Path(filepath).exists():
         return Path(filepath).stat().st_size
+    else:
+        msg = f"{filepath=} cannot be found."
+        logger.error(msg)
+        raise FileNotFoundError(msg)
+
+
+def md5_sum(filepath: str):
+    """
+    Get md5sum of a specific file on the local file system.
+
+    Parameters
+    ----------
+        filepath (string): The filepath
+
+    Returns
+    -------
+    The md5sum of the file.
+    """
+    if Path(filepath).exists():
+        with open(filepath, "rb") as f:
+            return hashlib.md5(f.read()).hexdigest()
     else:
         msg = f"{filepath=} cannot be found."
         logger.error(msg)