+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [semantic versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added a function in `io/input.py` called `file_size` to check a
   file size in a local drive.
+- Added a function in `io/input.py` called `md5_sum` to create md5
+  hash for an object in a local drive.
 
 ### Changed